scripts/pyninja/requirements.txt
pyYAML==5.1.1